add styling for quote section, add aditional syling to other sections
added the styling for the quote section as well as adjusted some of
the padding numbers for the other sections. I did this because I
thought what I had previously wasn't entirely accurate to how the
layout looked so I think this change will make the page look more like
the layout.